pyproject.toml
[project]
name = "fashion-detection-demo"
version = "1.0.0"
description = "Fashion detection web demo with FastAPI"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.35.0",
    "numpy<2.0.0",
]
